# Conflicts:
#	src/data/autoPart/AutoPart.txt
#	src/data/car/Car.txt
#	src/data/transaction/SaleTransaction.txt
#	src/data/user/UserAccount.txt
#	src/utils/menu/StatisticsMenu.java

+++ src/utils/menu/SaleTransactionMenu.java
@@ -1,6 +1,7 @@
 package utils.menu;
 
 import services.ServiceList;
+import transaction.SaleTransaction;
 import transaction.SaleTransactionList;
 import user.Manager;
 import user.Salesperson;
@@ -17,7 +18,7 @@ public class SaleTransactionMenu extends Menu {
         switch (currentUser) {
             case Manager m -> initializeMenu(MenuOption.MANAGER);
             case Salesperson s -> initializeMenu(MenuOption.SALESPERSON);
-            case null, default -> throw new IllegalArgumentException("Unsupported user type");
+            case null, default -> initializeMenu(null);
         }
     }
 
@@ -38,13 +39,14 @@ public class SaleTransactionMenu extends Menu {
                 menuItems.put(1, "Display all transactions by me");
 
 
-                menuActions.put(1, this::displayAllTransactions); // TODO: Display all transactions by salesperson
-                menuActions.put(2, this::searchTransactionById); // TODO: Search transactions by salesperson
-                menuActions.put(3, this::deleteTransactionWrapper); // TODO: Delete transactions by salesperson
-                menuActions.put(4, this::createTransactionWrapper); // TODO: Create transactions by salesperson
-                menuActions.put(5, this::updateTransactionWrapper); // TODO: Update transactions by salesperson
+                menuActions.put(1, this::displayAllTransactions);
+                menuActions.put(2, this::searchTransactionById);
+                menuActions.put(3, this::deleteTransactionWrapper);
+                menuActions.put(4, this::createTransactionWrapper);
+                menuActions.put(5, this::updateTransactionWrapper);
                 menuActions.put(0, this::exit);
             }
+            case null, default -> System.out.print("");
         }
         menuItems.put(2, "Search a transaction by ID");
         menuItems.put(3, "Delete a transaction");
@@ -87,25 +89,67 @@ public class SaleTransactionMenu extends Menu {
         Scanner input = new Scanner(System.in);
         System.out.println("Enter transaction ID: ");
         String transactionID = input.nextLine();
-        if (ServiceList.getServiceById(transactionID) != null) {
-            System.out.println("Transaction found!");
-            System.out.println(SaleTransactionList.getSaleTransactionById(transactionID).getFormattedSaleTransactionDetails());
+        User currentUser = UserSession.getCurrentUser();
+        SaleTransaction transaction = SaleTransactionList.getSaleTransactionById(transactionID);
+        if (currentUser.getRole() == User.ROLE.MANAGER){
+            if (transaction != null && !transaction.isDeleted()) {
+                System.out.println("Transaction found!");
+                System.out.println(transaction.getFormattedSaleTransactionDetails());
+            }
+            System.out.println("No transaction found with ID: " + transactionID);
+        } else if (currentUser.getRole() == User.ROLE.EMPLOYEE){
+            if (currentUser instanceof Salesperson){
+                if(transaction != null && !transaction.isDeleted()){
+                    if(transaction.getSalespersonId().equals(currentUser.getUserID())){
+                        System.out.println("Transaction found!");
+                        System.out.println(transaction.getFormattedSaleTransactionDetails());
+                    } else {
+                        System.out.println("You are not allow to access this transaction: " + transactionID);
+                    }
+                } else {
+                    System.out.println("No transaction found with ID: " + transactionID);
+                }
+            }
         }
-        System.out.println("No transaction found with ID: " + transactionID);
     }
 
     private void deleteTransaction() throws Exception {
         System.out.println("Deleting a transaction...");
-        SaleTransactionList.deleteSaleTransaction();
+        User currentUser = UserSession.getCurrentUser();
+        if (currentUser.getRole() == User.ROLE.MANAGER){
+            SaleTransactionList.deleteSaleTransaction();
+        } else if (currentUser.getRole() == User.ROLE.EMPLOYEE){
+            if (currentUser instanceof Salesperson){
+
+            }
+        }
     }
 
     private void createNewTransaction() throws Exception {
         System.out.println("Creating a new transaction...");
         if (UserSession.getCurrentUser().getRole() == User.ROLE.MANAGER) {
             Scanner input = new Scanner(System.in);
-            System.out.println("Enter saleperson ID: ");
-            String salepersonID = input.nextLine();
-            SaleTransactionList.addSaleTransaction(salepersonID);
+            User saleperson = null;
+            while (saleperson == null) {
+                System.out.println("Enter saleperson ID: ");
+                String saleprsonId = input.nextLine();
+
+                String finalSalepersonId = saleprsonId;
+                saleperson = UserMenu.getUserList().stream()
+                        .filter(u -> u.getUserID().equals(finalSalepersonId))
+                        .findFirst()
+                        .orElse(null);
+
+                if (saleperson == null) {
+                    System.out.print("Invalid client ID. Please press enter to retype or 'quit' to exit: ");
+                    saleprsonId = input.nextLine();
+                    if (saleprsonId.equalsIgnoreCase("quit")) {
+                        System.out.println("Exiting..");
+                        return;
+                    }
+                }
+            }
+            SaleTransactionList.addSaleTransaction(saleperson.getUserID());
         } else {
             ServiceList.addService(UserSession.getCurrentUser().getUserName());
         }
